[#9] Fix a race condition when creating monitoring items.
The current API requires MonitoringMode.Sampling be used when creating items in order to allow callers a chance to set consumers and do other bookkeeping before values arrive, and then MonitoringMode.Reporting to be set once ready.

The new method allows callers to provide a callback that will be run with a guarantee that it happens before values can arrive.