#26: DefaultCertificateValidator may leak resources
This change allows one to properly close the validator, cleaning up the
internal thread and file system watcher.

It also adds a new constructor which allows to construct the instance
providing all necessary directories and allows one to disable writing
the rejected certificates.

A test case was also added to test the shutdown.

Signed-off-by: Jens Reimann <jreimann@redhat.com>