Make client verification of server certificates mandatory
Previously, it was up to implementations to do verification and
validation using the certificate in the chosen endpoint. Going forward,
the stack will validate and verify the first time it receives the
certificate in an asymmetric security header.

An insecure validator implementation that warns but trusts all
certificates is configured by default to ease the transition from a
backwards compatibility standpoint.