Refactor UaNode management (#279)
Biggest changes:

NodeMap becomes NodeManager, which no longer pretends to be a Map.

UaNode References are stored in the NodeManager, not the UaNode itself. This
allows forward and/or inverse references to be derived and generally makes
linking UaNodes together easier, especially when they belong to different
namespaces. For example, adding a UaNode underneath the Objects folder only
requires the user to add an inverse reference pointing to Objects. The forward
reference from Objects will be derived when needed.

This also lays the groundwork for an EventFactory that can instatiate fully
formed Event Object instances without those instances becoming part of the
server's visible address space by using a special UaNodeManager implementation.