Fix a "deadlock" in SessionFsm during Closing state
This happens because Closing state had 3 transitions that weren't
asynchronously calling fireEvent, causing it to block indefinitely
trying to obtain the write lock.

Further, change the implementation of SessionFsm#fireEvent so it does
not block if called while the write lock is already held; instead just
fire the event asynchronously for the caller.

fixes #309