Fix corner case in sequence number tracking that skipped a republish
One or more keep-alive notifications followed by a data change
notification that got lost or dropped was not being detected because the
last seen sequence number was being set for keep-alive notifications as
well as data change notifications.

A series like this would be miss the republish:

[seq=1] data change
[seq=2] keep-alive
[seq=2] keep-alive
[seq=2] data change (packet dropped)
[seq=3] keep-alive or data change

because the keep-alives with seq=2 were marked as last seen, meaning
when seq=3 arrived the sequence appeared intact.