Check for fragments larger than requested when reading DataTypeDictionary
If a server doesn't properly support reading the index range it may
just return the entire contents of the dictionary in a single call.
This resulted in an a never ending series of calls to keep reading the
next fragment, waiting for a partial fragment that would never arrive
to terminate.

An edge case where the dictionary size is exactly equal to the
fragment size still exists. In this case we must hope the server
properly terminates the subsequent request with something like
Bad_IndexRangeNoData or else the infinite loop could still happen.