Don't allow connect persistence to be configured
The SessionFsm relies on being able to disconnect and reconnect the
underlying channel when the keep alive fails. If connect persistence is
disabled the ChannelFsm will go to NotConnected rather than its
reconnecting states and the promise that once connected it will try
to remain connected will be broken. Further, the session re-activation
depends on the exponential back-off in delay between reconnect attempts
at the ChannelFsm layer. Without this re-activation attempts will occur
in a tight loop.