Per-Namespace NodeManagers collated by server NodeManager (#394)
Require Namespace's to provide a local NodeManager used to store Nodes and References belonging to that Namespace. The server-wide NodeManager now just aggregates/collates the various sub-NodeManagers for each Namespace. 

When adding a reference where both Nodes are in the same Namespace the local NodeManager can be used, but if one of the Nodes belongs to another Namespace then the Reference should be added to the server NodeManager so both local NodeManagers end up having the appropriate Reference.