Generate new UA namespace nodes (#468)
ObjectNode and VariableNodes now follow a different naming scheme.

Previously they assumed the type node they were modelling ended with a
"Type" suffix and tried to replace that with "Node". It turns out that
other information models do not follow this naming practice, so now the
code generator just appends "Node" to the BrowseName of the type node.

QualifiedProperty has also been augmented to accept an ExpandedNodeId 
for the `dataType` parameter and the type nodes have been generated with 
properties that use ExpandedNodeId.