Implement new AttributeFilter concept (#476)
AttributeFilters are a replacement for the AttributeDelegate concept that resolve some of the inconsistencies with delegates, namely, that calling the getter or setter for any given attribute did not use the registered delegate and so for any attribute whose value was provided by a delegate you could have different values depending on whether you retrieved that value via readAttribute/writeAttribute or its getter/setter.

AttributeFilters can also be chained an a more natural way. Filters are added to front or back of a filter chain, which can be thought of as a linked list of filters. When an attribute get or set happens, each filter in the chain gets a chance to process the operation and choose if/when to call the next filter in the chain.

All filter chains, including the default empty chain, have an implicit filter at the end that simply reads or writes to the backing field for the attribute being get or set.