Pass messages to next the handler using ctx.fireChannelRead()
This prevents sequence numbers from being processed out of order when a
two or more messages, starting with an OPN for a secure channel renewal,
arrive at the same time and are all ready to be decoded.

Implement `decode` consistently among all the handlers, allowing Netty
to be responsible for looping until no more messages can be read.

fixes #483