Notify subscription transfer failed after *any* failure
This prevents the FSM from getting stuck in a reconnect "loop" when a
defective server doesn't respond with one of the expected StatusCodes or
doesn't respond with a ServiceFault at all, e.g. responding instead with
a channel layer Error message instead.

The main consequence of this is that subscription transfer will only be
attempted once after re-establishing a session, leaving a small possibility
that some legitimate type of failure, such as a connection interruption,
occurring as the subscriptions are being transferred results in the
subscriptions being subsequently re-created when it may have been possible
to transfer them.

fixes #722