Remove check preventing Variant construction with DataValue (#806)
OPC UA 1.04 further clarifies that it is legal to construct a Variant
containing a DataValue for use in e.g. method arguments or PubSub but it is
not legal for a DataValue to contain such a Variant. Without knowledge of
where the Variant will be used it does not make sense to restrict construction
of such a Variant.

fixes #805