Fix NPE generating struct with enum field
When generating the BSD XML for a struct that has an enum type field
assume the namespace URI is that of the current dictionary. Structs
are not currently defined with enough extra information to look up the
namespace URI; consequently a struct definition may not reference an
enum type defined in another dictionary.

fixes #964