Add AttributeFilter `::readAttribute` and `::writeAttribute`
These are now called by AttributeReader and AttributeWriter when responding to Read and Write service calls, and allow for filter implementations that can indicate a failure to get or set an attribute by throwing UaException, as opposed to the `getAttribute` and `setAttribute` methods that must not throw.

Additionally, filter contexts are simplified, removing the separate GetAttributeContext and SetAttributeContext in favor of just AttributeFilterContext.